experiments/advanced-rag-vision-cohere/pyproject.toml
[project]
name = "vision-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.15.0",
    "fire>=0.7.0",
    "google-genai>=1.12.1",
    "numpy>=2.2.5",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]
